fix(schema): GREEN — widen the TiB capacity fields from int to float
usable_capacity_tib and int_client_store_tib become float. A usable
capacity is rarely a whole number of TiB once RAID/EC overhead is taken
out — XSKY's AIMesh systems report 111.78 — and rounding a submitter's
stated figure to satisfy the schema is not ours to do.

  * schema_validator.py — int -> float, Optional[int] -> Optional[float]
  * schema.yaml — int( min=1 ) -> num( min=1 ) in the reference schema
  * conftest.py / test_reportgen_sut_derived.py — helper annotations

ge=1 is deliberately unchanged: the widening is the type only, not the
accepted range, so sub-1-TiB values are still rejected.

Whole numbers keep validating. The model is lax — StrictModel sets
extra="forbid", not strict=True — so an int coerces to float, and every
example_*.yaml template still passes.

No results-table output changes: report_generator.py reads the raw YAML
via solution.get() and never goes through the pydantic model, so an
integer capacity still renders as 8192, not 8192.0.

Full CI-equivalent sweep green: tests 3146 passed / 1 skipped,
mlpstorage_py/tests 911, kv_cache_benchmark 238, vdb_benchmark 228.

+++ mlpstorage_py/system_description/schema_validator.py
@@ -311,9 +311,16 @@ class Solution(StrictModel):
     # Availability and Integrated Client Storage (TiB) columns. The first two
     # are required; int_client_store_tib is optional — no v3.0 submitter is
     # expected to use it, the column survives for v2.0 parity.
-    usable_capacity_tib:  int          = Field(ge=1)
+    #
+    # Both capacities are float, not int: a usable capacity is rarely a whole
+    # number of TiB once RAID/EC overhead is taken out (XSKY's AIMesh reports
+    # 111.78), and rounding a submitter's stated figure to satisfy the schema
+    # is not ours to do. ge=1 is unchanged by the widening — sub-1-TiB is
+    # still out of range. Whole numbers keep validating: the model is lax
+    # (extra="forbid" only, not strict=True), so an int coerces to float.
+    usable_capacity_tib:  float          = Field(ge=1)
     availability:         Availability
-    int_client_store_tib: Optional[int] = Field(default=None, ge=1)
+    int_client_store_tib: Optional[float] = Field(default=None, ge=1)
 
 
 class SystemUnderTest(StrictModel):
